chore: 폴더 구조 및 파일명 변경
