+++ src/App.tsx
@@ -3,15 +3,15 @@ import styled from 'styled-components';
 import reset from 'styled-reset';
 import { createGlobalStyle, ThemeProvider } from 'styled-components';
 import theme from './styles/theme';
-import Header from './components/layouts/Header';
-import Footer from './components/layouts/Footer';
+import Header from './components/layouts/common/Header';
+import Footer from './components/layouts/common/Footer';
 import { Outlet } from 'react-router-dom';
 import { User } from './utils/interfaces';
 
 const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
+	display: flex;
+	flex-direction: column;
+	min-height: 100vh;
 `;
 
 const GlobalStyle = createGlobalStyle`
@@ -37,40 +37,40 @@ text-decoration: none;
 `;
 
 const ContentsWrapper = styled.div`
-  flex: 1;
+	flex: 1;
 `;
 
 function App() {
-  const [user, setUser] = useState<User | null>(null);
+	const [user, setUser] = useState<User | null>(null);
 
-  const getUser = async () => {
-    try {
-      const res = await fetch('https://bootcamp-api.codeit.kr/api/sample/user');
-      const data = await res.json();
-      return data;
-    } catch (error) {
-      console.log(error);
-    }
-  };
+	const getUser = async () => {
+		try {
+			const res = await fetch('https://bootcamp-api.codeit.kr/api/sample/user');
+			const data = await res.json();
+			return data;
+		} catch (error) {
+			console.log(error);
+		}
+	};
 
-  useEffect(() => {
-    getUser().then((data) => {
-      setUser(data);
-    });
-  }, []);
+	useEffect(() => {
+		getUser().then((data) => {
+			setUser(data);
+		});
+	}, []);
 
-  return (
-    <Container>
-      <ThemeProvider theme={theme}>
-        <GlobalStyle />
-        <Header email={user?.email} profileImageSource={user?.profileImageSource} />
-        <ContentsWrapper>
-          <Outlet />
-        </ContentsWrapper>
-        <Footer />
-      </ThemeProvider>
-    </Container>
-  );
+	return (
+		<Container>
+			<ThemeProvider theme={theme}>
+				<GlobalStyle />
+				<Header email={user?.email} profileImageSource={user?.profileImageSource} />
+				<ContentsWrapper>
+					<Outlet />
+				</ContentsWrapper>
+				<Footer />
+			</ThemeProvider>
+		</Container>
+	);
 }
 
 export default App;
