+++ src/pages/Shared.tsx
@@ -1,18 +1,18 @@
 import styled from 'styled-components';
-import ProfileSection from '../components/layouts/ProfileSection';
-import ListSection from '../components/layouts/ListSection';
+import ProfileSection from '../components/layouts/shared/ProfileSection';
+import SharedListSection from '../components/layouts/shared/SharedListSection';
 
 const Container = styled.div`
-  width: 100%;
+	width: 100%;
 `;
 
 function Shared() {
-  return (
-    <Container>
-      <ProfileSection />
-      <ListSection />
-    </Container>
-  );
+	return (
+		<Container>
+			<ProfileSection />
+			<SharedListSection />
+		</Container>
+	);
 }
 
 export default Shared;
